requirements.txt
Flask==2.0.2
numpy==1.21.4
google-cloud-aiplatform
google-cloud-storage
gunicorn==19.7.1